Refactor: Removed unused code and moved maps to common utils.
- Commented and unsed code fragments removed.
- Refactored all map components to import maps from the common utils
  folder.
- Updated map to latest map from UBOS.
- Regression noticed when changing tabs because of the map size
