+++ src/views/Dashboard/components/CoveragePanel/components/CoverageByMonth/AnnualizedCoverageForAntigensChart/index.js
@@ -14,7 +14,6 @@ import {
 } from "./chart";
 
 // Utils
-
 import { generateChartTitle } from "../../../../../../../common/utils/utils";
 
 const CoverageForAntigensChart = props => {
