+++ src/views/Dashboard/components/CoveragePanel/components/CoverageByMonth/AnnualizedCoverageForVaccineMap/chart.js
@@ -17,7 +17,8 @@ import {
 
 //  Map
 
-let ugandaMap = require("./map.json");
+const ugandaMap = require("../../../../../../../common/maps/map.json");
+const ugandaMap2 = require("../../../../../../../common/maps/map2.json");
 
 export const vaccineAnnualizedCoverage = (
   data,
@@ -49,7 +50,7 @@ export const vaccineAnnualizedCoverage = (
 
     // Data District name is in format `District District`.
     // We only want the mane minus the word district at the end for our map
-    const districtName = district.replace(/ District/g, "");
+    const districtName = district.replace(/ District/g, "").toUpperCase();
 
     mapData.push([districtName, coverageRate]);
   }
@@ -57,7 +58,7 @@ export const vaccineAnnualizedCoverage = (
   return {
     ...commonChartOptions,
     chart: {
-      map: ugandaMap,
+      map: ugandaMap2,
       height: 80 + "%"
     },
 
@@ -102,7 +103,7 @@ export const vaccineAnnualizedCoverage = (
       formatter: function() {
         return (
           "<b><u>" +
-          this.point.properties.name +
+          this.point.properties.DName2018 +
           "</u></b><br/><br/>" +
           +Highcharts.numberFormat(this.point.value, 1) +
           " %"
@@ -113,8 +114,8 @@ export const vaccineAnnualizedCoverage = (
     series: [
       {
         data: mapData,
-        keys: ["name", "value"],
-        joinBy: "name",
+        keys: ["DName2018", "value"],
+        joinBy: "DName2018",
         borderColor: "grey",
         borderWidth: 0.5,
         states: {
